simple_lmk: Introduce simple low memory killer for Android
This is a simple replacement for the Android Low Memory Killer driver that
simply targets a predefined number of pages to be freed every time a memory
allocation attempt is about to utilize direct reclaim. Processes are killed
according to the priorities that Android gives them; in other words, the
least important processes are always killed first.

Simple LMK memory reclaim is triggered in the page-alloc slow path after
compaction has failed. When the Simple LMK memory reclaim process is
entered, memory is reclaimed in an infinite loop until a normal allocation
succeeds. Repeatedly looping on compaction is intentionally avoided since
compaction is expensive and adds to the overall memory allocation latency;
after compaction hasn't succeeded once for the allocation, it isn't
repeated for the same allocation.

Signed-off-by: Sultan Alsawaf <sultanxda@gmail.com>